feat: add auth to users

+++ src/users/users.controller.ts
@@ -8,15 +8,20 @@ import {
   Delete,
   UseInterceptors,
   HttpCode,
+  UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { ApiOperation, ApiParam, ApiResponse, ApiTags } from '@nestjs/swagger';
 
+import { ValidatePayloadExistsPipe } from 'src/shared/pipes/validate-payload-exist.pipe';
+import { ConflictInterceptor } from 'src/shared/interceptors/conflict.interceptor';
+import { GetUser } from 'src/auth/decorators/get-user.decorator';
+import { JwtGuard } from 'src/auth/guards/jwt.guard';
+import { AuthUser } from 'src/auth/auth-user.type';
 import { User } from './entities/user.entity';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersService } from './users.service';
-import { ValidatePayloadExistsPipe } from 'src/shared/pipes/validate-payload-exist.pipe';
-import { ConflictInterceptor } from 'src/shared/interceptors/conflict.interceptor';
 
 @Controller('users')
 @ApiTags('users')
@@ -47,6 +52,7 @@ export class UsersController {
   }
 
   @Patch(':id')
+  @UseGuards(JwtGuard)
   @ApiOperation({ summary: 'Update user data' })
   @ApiParam({ name: 'id', description: 'User id' })
   @UseInterceptors(ConflictInterceptor)
@@ -55,15 +61,28 @@ export class UsersController {
     @Param('id') id: string,
     @Body(new ValidatePayloadExistsPipe())
     updateUserDto: UpdateUserDto,
+    @GetUser() user: AuthUser,
   ): Promise<void> {
+    if (id !== user.id) {
+      throw new UnauthorizedException();
+    }
+
     return this.usersService.updateUser(id, updateUserDto);
   }
 
   @Delete(':id')
+  @UseGuards(JwtGuard)
   @ApiOperation({ summary: 'Delete user' })
   @ApiParam({ name: 'id', description: 'User id' })
   @HttpCode(204)
-  async removeUser(@Param() id: string): Promise<void> {
+  async removeUser(
+    @Param() id: string,
+    @GetUser() user: AuthUser,
+  ): Promise<void> {
+    if (id !== user.id) {
+      throw new UnauthorizedException();
+    }
+
     return this.usersService.removeUser(id);
   }
 }
